Displaying a list with colored cubes.
I was able to get the list to display with colored cubes next to the list items.
The data object currently stores the RGB values as strings, and converts into numbers.
Unsure if this will remain the case moving forwards.
There's a typing error that prints to the console when running, but it seems to dipislay.

Issue #:3